refactor: separate out windows implementation into its own files based on the version
