+++ job/window_v2_test.go
@@ -256,232 +256,3 @@ func TestWindowV2(t *testing.T) {
 		})
 	})
 }
-
-func TestWindowV1(t *testing.T) {
-	t.Run("Validate", func(t *testing.T) {
-		t.Run("should not throw error for window size which is not a positive or an instant time duration", func(t *testing.T) {
-			validWindowConfigs := []string{"24h", "2h45m", "60s", "45m24h", "", "0"}
-			for _, config := range validWindowConfigs {
-				window := job.WindowV1{Size: config}
-				err := window.Validate()
-				assert.Nil(t, err, fmt.Sprintf("failed for : %s", config))
-			}
-		})
-		t.Run("should throw error for window size which is not a valid time duration", func(t *testing.T) {
-			inValidWindowConfigs := []string{"1M", "45M24h", "60S", "60", "2d", "-24h"}
-			for _, config := range inValidWindowConfigs {
-				window := job.WindowV1{Size: config}
-				err := window.Validate()
-				assert.NotNil(t, err, fmt.Sprintf("failed for %s", config))
-			}
-		})
-		t.Run("should not throw error for window offset which is a valid time duration", func(t *testing.T) {
-			validOffsetConfigs := []string{"24h", "2h45m", "60s", "45m24h", "0", ""}
-			for _, config := range validOffsetConfigs {
-				window := job.WindowV1{Offset: config}
-				err := window.Validate()
-				assert.Nil(t, err, fmt.Sprintf("failed for : %s", config))
-			}
-		})
-		t.Run("should throw error for window offset which is not a valid time duration", func(t *testing.T) {
-			inValidOffsetConfigs := []string{"1M", "45M24h", "60S", "60"}
-			for _, config := range inValidOffsetConfigs {
-				window := job.WindowV1{Offset: config}
-				err := window.Validate()
-				assert.NotNil(t, err, fmt.Sprintf("failed for %s", config))
-			}
-		})
-		t.Run("should not throw error for valid window truncate configs", func(t *testing.T) {
-			validTruncateConfigs := []string{"h", "d", "w", "m", "M", ""}
-			for _, config := range validTruncateConfigs {
-				window := job.WindowV1{TruncateTo: config}
-				err := window.Validate()
-				assert.Nil(t, err, fmt.Sprintf("failed for : %s", config))
-			}
-		})
-		t.Run("should throw error for window truncate when it is not a valid time duration", func(t *testing.T) {
-			inValidTruncateConfigs := []string{"s", "a", "ms"}
-			for _, config := range inValidTruncateConfigs {
-				window := job.WindowV1{TruncateTo: config}
-				err := window.Validate()
-				assert.NotNil(t, err, fmt.Sprintf("failed for %s", config))
-			}
-		})
-	})
-	t.Run("GetTimeRange", func(t *testing.T) {
-		t.Run("should provide start and end time window for the given schedule", func(t *testing.T) {
-			cases := []struct {
-				Scenario          string
-				ScheduleTime      time.Time
-				Size              string
-				Offset            string
-				TruncateTo        string
-				ExpectedStartTime time.Time
-				ExpectedEndTime   time.Time
-			}{
-				{
-					Scenario:          "should truncate to the previous hour on truncate to hour",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "h",
-					ExpectedStartTime: time.Date(2022, 07, 04, 02, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 02, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should truncate to the previous day on truncate to day",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 04, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should truncate to the coming sunday on truncate to week",
-					ScheduleTime:      time.Date(2020, 07, 10, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "w",
-					ExpectedStartTime: time.Date(2020, 07, 11, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2020, 07, 12, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should truncate to the start of the next month on truncate to month",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "m",
-					ExpectedStartTime: time.Date(2022, 8, 1, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 8, 1, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should be able to get a single month window if size is greater than 30h",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "32h",
-					Offset:            "0",
-					TruncateTo:        "m",
-					ExpectedStartTime: time.Date(2022, 7, 1, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 7, 31, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should not truncate to the previous hour if time is already truncated",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 00, 00, 00, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "h",
-					ExpectedStartTime: time.Date(2022, 07, 04, 02, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 02, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should not truncate to the previous day if time is already truncated",
-					ScheduleTime:      time.Date(2022, 07, 05, 00, 0, 0, 0, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 04, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should truncate to next week sunday if the time is already on exact sunday on truncate to week",
-					ScheduleTime:      time.Date(2020, 07, 12, 00, 0, 0, 0, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "w",
-					ExpectedStartTime: time.Date(2020, 07, 18, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2020, 07, 19, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should truncate to the end of month if time is already on beginning of month on truncate to month",
-					ScheduleTime:      time.Date(2022, 07, 01, 0, 0, 0, 0, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "m",
-					ExpectedStartTime: time.Date(2022, 8, 1, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 8, 1, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should truncate to daily if truncate is empty",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "0",
-					TruncateTo:        "",
-					ExpectedStartTime: time.Date(2022, 07, 04, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should provide window for the configured size in hours and minutes",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h30m",
-					Offset:            "0",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 03, 23, 30, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should provide window for the configured size in minutes",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "60m",
-					Offset:            "0",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 04, 23, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should provide 24 hour window if size is not configured",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "",
-					Offset:            "0",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 04, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should provide monthly window, with monthly truncation and 30h size is one month",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "62h",
-					Offset:            "0",
-					TruncateTo:        "m",
-					ExpectedStartTime: time.Date(2022, 6, 01, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 8, 01, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should shift window forward on positive offset",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "24h",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 06, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should shift window backward on negative offset",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "-24h",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 03, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 04, 0, 0, 0, 0, time.UTC),
-				},
-				{
-					Scenario:          "should not shift window when offset is not configured",
-					ScheduleTime:      time.Date(2022, 07, 05, 02, 10, 10, 10, time.UTC),
-					Size:              "24h",
-					Offset:            "",
-					TruncateTo:        "d",
-					ExpectedStartTime: time.Date(2022, 07, 04, 0, 0, 0, 0, time.UTC),
-					ExpectedEndTime:   time.Date(2022, 07, 05, 0, 0, 0, 0, time.UTC),
-				},
-			}
-			for _, sc := range cases {
-				w := job.WindowV1{Size: sc.Size, Offset: sc.Offset, TruncateTo: sc.TruncateTo}
-				err := w.Validate()
-				assert.Nil(t, err, sc.Scenario)
-				actualStartTime, actualEndTime, err := w.GetTimeRange(sc.ScheduleTime)
-				assert.Nil(t, err, sc.Scenario)
-				assert.Equal(t, sc.ExpectedStartTime.String(), actualStartTime.String(), sc.Scenario)
-				assert.Equal(t, sc.ExpectedEndTime.String(), actualEndTime.String(), sc.Scenario)
-			}
-		})
-	})
-}
